refactor location
