+++ internal/app/feed/feed.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/SafetyCulture/safetyculture-exporter/internal/app/api"
-	"github.com/SafetyCulture/safetyculture-exporter/internal/app/config"
 )
 
 // Feed is an interface to a data feed. It provides methods to export the data to an exporter
@@ -43,32 +42,6 @@ type Exporter interface {
 
 	SupportsUpsert() bool
 	ParameterLimit() int
-}
-
-// SafetyCultureExporter defines the basic action in regard to the exporter
-type SafetyCultureExporter interface {
-	// CreateSchemas will generate the schema for SafetyCulture feeds, without downloading data
-	CreateSchemas(exporter Exporter) error
-	// ExportFeeds will export SafetyCulture feeds and Sheqsy feeds
-	ExportFeeds(exporter Exporter) error
-}
-
-type ExporterApp struct {
-	exportConfig    *config.ExportConfig
-	apiClient       *api.Client
-	apiConfig       *config.ApiConfig
-	sheqsyApiClient *api.Client
-	sheqsyApiConfig *config.SheqsyApiConfig
-}
-
-func NewExporterApp(scApiClient *api.Client, sheqsyApiClient *api.Client, cfg *config.ConfigurationOptions) *ExporterApp {
-	return &ExporterApp{
-		exportConfig:    cfg.ExportConfig,
-		apiClient:       scApiClient,
-		apiConfig:       cfg.ApiConfig,
-		sheqsyApiClient: sheqsyApiClient,
-		sheqsyApiConfig: cfg.SheqsyApiConfig,
-	}
 }
 
 // DeduplicateList a list of T type and maintains the latest value
